refactor(gsd): extract auto-worktree runtime helpers
